changed method signature of size() method and added comments.

+++ ptolemy/domains/de/kernel/DEEventQueue.java
@@ -82,7 +82,11 @@ public interface DEEventQueue extends Debuggable {
      */
     public void put(DEEvent event) throws IllegalActionException;
 
-    public double size();
+    /** Return the size of the event queue.  
+     *  @return The size of the event queue.
+     */
+    public int size();
+    
     /** Dequeue the earliest event in this event queue.
      *  @return The DEEvent object associated with the earliest event in
      *    the queue.
